+++ macaw-aarch32/ChangeLog.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Added support for symbolically executing system calls
+- Added shorthands for `r0` through `r14` to `ARMReg.hs`
 
 ### Breaking Changes
 
@@ -12,3 +13,4 @@
 
   The terminators are gone, replaced by additional function forms. Note that these are compiled away during the translation into crucible.
 
+- Renamed `arm_LR` to `lr`